rename cpu/gpu modules
