+++ src/fftvis/__init__.py
@@ -2,12 +2,12 @@
 
 # Import key components for beams
 from .core.beams import BeamEvaluator
-from .cpu.cpu_beams import CPUBeamEvaluator
+from .cpu.beams import CPUBeamEvaluator
 from .wrapper import create_beam_evaluator
 
 # Import simulation functionality
 from .core.simulate import SimulationEngine
-from .cpu.cpu_simulate import CPUSimulationEngine
+from .cpu.simulate import CPUSimulationEngine
 from .wrapper import create_simulation_engine, simulate_vis
 
 # Import utility modules
