+++ src/fftvis/cpu/__init__.py
@@ -1,5 +1,5 @@
 """CPU-specific implementations for fftvis."""
 
-from .cpu_beams import CPUBeamEvaluator
-from .cpu_simulate import CPUSimulationEngine
-from .cpu_nufft import cpu_nufft2d, cpu_nufft3d
+from .beams import CPUBeamEvaluator
+from .simulate import CPUSimulationEngine
+from .nufft import cpu_nufft2d, cpu_nufft3d, cpu_nufft2d_type1
