+++ src/fftvis/cpu/beams.py
