+++ src/fftvis/gpu/__init__.py
@@ -1,5 +1,5 @@
 """GPU-specific implementations for fftvis."""
 
-from .gpu_beams import GPUBeamEvaluator
-from .gpu_simulate import GPUSimulationEngine
-from .gpu_nufft import gpu_nufft2d, gpu_nufft3d
+from .beams import GPUBeamEvaluator
+from .simulate import GPUSimulationEngine
+from .nufft import gpu_nufft2d, gpu_nufft3d
