+++ src/fftvis/gpu/beams.py
